+++ docs/changelog.md
@@ -14,6 +14,7 @@
 - Added `hybridToggle` and `AxeHybridToggle` patch types that normally work as `toggle` but work as `momentary` if the release of a button press takes longer than 250ms; bank switch secondary functions take precedence over momentary behavior of a hybridToggle. [#24](../../../issues/24)
 - Added support for exclusive group patches via `groupId` attribute on patch definition.  Activating one patch in a group will deactivate any other previously active patches in the group (a way to support radio button exclusivity behavior  independent of bank definitions). [#37](../../../issues/37)
 - Fixed unintended application window deactivation after using Ctrl+O to load a new autogrid config. [#33](../../../issues/33)
+- In Bank Direct mode, added listing of all banks in bank definition order to the main display. [#38](../../../issues/38)
 - Removed backwards compatible support for user defined bank and patch numbers; all bank and patch references are now only by name.  Removed display of bank and patch numbers from the main display.
 
 #### 2024.02.05